refactor: move interface.tsx
